{{cookiecutter.project_slug}}/django_hello_world/zappa_django_hello_world/settings/storage.py
from django.conf import settings
from storages.backends.s3boto import S3BotoStorage


class MediaStorage(S3BotoStorage):
    location = settings.MEDIAFILES_LOCATION


class StaticStorage(S3BotoStorage):
    location = settings.STATICFILES_LOCATION
