{{cookiecutter.project_slug}}/django_hello_world/requirements.txt
django
django-storages
boto
boto3
zappa
